bot/pyproject.toml
[project]
name = "medvychet-bot"
version = "0.1.0"
description = "MedВычет Telegram bot"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=21.6",
    "httpx>=0.27.0",
    "pydantic-settings>=2.6.0",
    "sentry-sdk>=2.18.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
